==> 4_CNN.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load in 

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the "../input/" directory.
# For example, running this (by clicking run or pressing Shift+Enter) will list the files in the input directory

import os
#print(os.listdir("../input"))

# Any results you write to the current directory are saved as output.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

df = pd.read_csv("processed_data.csv", index_col=0)

df.head()

X = df["text"].values
y = df.drop(columns="text").values 

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

from keras.preprocessing import sequence
from keras.models import Model, Input
from keras.layers import Dense, Embedding, GlobalMaxPooling1D, Conv1D, Dropout
from keras.preprocessing.text import Tokenizer
from keras.optimizers import Adam

max_features = 20000  # number of words we want to keep
maxlen = 300  # max length of the comments in the model
batch_size = 64  # batch size for the model
embedding_dims = 20  # dimension of the hidden variable, i.e. the embedding dimension

tokenizer = Tokenizer(num_words=max_features)
tokenizer.fit_on_texts(list(X_train) + list(X_test))
x_train = tokenizer.texts_to_sequences(X_train)
x_test = tokenizer.texts_to_sequences(X_test)
print(len(x_train), 'train sequences')
print(len(x_test), 'test sequences')
print('Average train sequence length: {}'.format(np.mean(list(map(len, x_train)), dtype=int)))
print('Average test sequence length: {}'.format(np.mean(list(map(len, x_test)), dtype=int)))

x_train = sequence.pad_sequences(x_train, maxlen=maxlen)
x_test = sequence.pad_sequences(x_test, maxlen=maxlen)
print('x_train shape:', x_train.shape)
print('x_test shape:', x_test.shape)

comment_emb.shape

x_train[0]

comment_input = Input((maxlen,))

# we start off with an efficient embedding layer which maps
# our vocab indices into embedding_dims dimensions
comment_emb = Embedding(max_features, 
                        embedding_dims, 
                        input_length=maxlen)(comment_input)
cnn1 = Conv1D(filters=50, kernel_size=5, activation='relu')(comment_emb)
# we add a GlobalMaxPooling1D, which will extract features from the embeddings
# of all words in the comment
h = GlobalMaxPooling1D()(cnn1)

# We project onto a six-unit output layer, and squash it with a sigmoid:
fc1 = Dense(200, activation='relu')(h)
d1 = Dropout(0.4)(fc1)
output = Dense(9, activation='sigmoid')(d1)

model = Model(inputs=comment_input, outputs=output)

model.compile(loss='binary_crossentropy',
              optimizer=Adam(0.01),
              metrics=['accuracy', 
                      'categorical_accuracy'])


history = model.fit(x_train, y_train, batch_size=batch_size, epochs=10, validation_split=0.1)


sample = 9
print(y_test[sample])
print(x_test.shape)
print(np.array(model.predict(x_test)[sample]>0.5).astype(int))
y_pred = (model.predict(x_test)>0.5).astype(int)

from sklearn.metrics import accuracy_score
import keras 

keras.backend.print_tensor(keras.metrics.top_k_categorical_accuracy(y_test, y_pred, k = 2))
